backend/requirements.txt
# AdEngineAI — Python Agent Sidecar
# Install: pip install -r requirements.txt
# Then:    playwright install chromium

# --- Web Framework ---
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1

# --- LLM — Development ---
groq==0.11.0

# --- LLM — Production ---
anthropic==0.34.2
openai==1.51.0

# --- AI Frameworks ---
langgraph==0.2.28
crewai==0.63.0
pyautogen==0.2.35

# --- Storage — Development ---
cloudinary==1.41.0

# --- Storage — Production ---
boto3==1.35.0

# --- Scraping ---
playwright==1.47.0
beautifulsoup4==4.12.3
lxml==5.3.0
httpx==0.27.2

# --- Data ---
pydantic==2.9.2

# --- Testing ---
pytest==8.3.3
pytest-asyncio==0.24.0

# AdEngineAI — Full Backend Dependencies
# Install: pip install -r requirements.txt

# ─── Web Framework ──────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic[email]==2.9.2

# ─── Database ───────────────────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0             # async PostgreSQL driver
alembic==1.13.3             # database migrations

# ─── Auth & Security ────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0   # JWT
passlib[bcrypt]==1.7.4             # password hashing
bcrypt==4.2.0

# ─── Redis ──────────────────────────────────────────────────────────────────
redis[asyncio]==5.1.0
hiredis==3.0.0              # faster Redis parser

# ─── Billing ────────────────────────────────────────────────────────────────
stripe==11.1.0

# ─── LLM ────────────────────────────────────────────────────────────────────
groq==0.11.0
anthropic==0.34.2
openai==1.51.0

# ─── AI Frameworks ──────────────────────────────────────────────────────────
langgraph==0.2.28
crewai==0.63.0
pyautogen==0.2.35

# ─── Storage ────────────────────────────────────────────────────────────────
cloudinary==1.41.0
boto3==1.35.0

# ─── Scraping ───────────────────────────────────────────────────────────────
playwright==1.47.0
beautifulsoup4==4.12.3
lxml==5.3.0
httpx==0.27.2

# ─── Video / Voice ──────────────────────────────────────────────────────────
elevenlabs==1.50.0

# ─── Testing ────────────────────────────────────────────────────────────────
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2               # for FastAPI test client

bcrypt==4.0.1

ffmpeg-python
httpx